fix(d): 补上条件最优性表的生成与论文中的表环境；RUF046
- report.py: _conditional_table 生成 tab_<tag>_conditional（值/已证明下界/状态/用时），
  阶段未传 conditional_time_limits 时为空操作。
- 04_models.tex: 插入 tab:q2conditional 的表环境，表题写明下界只对受限问题成立。

+++ pipelines/d/stages.py
@@ -94,7 +94,7 @@ def _max_independent_set(
     solver.parameters.max_time_in_seconds = float(time_limit)
     status = solver.Solve(cp)
     return {
-        "value": int(round(solver.ObjectiveValue())) if status in (cp_model.OPTIMAL, cp_model.FEASIBLE) else 0,
+        "value": round(solver.ObjectiveValue()) if status in (cp_model.OPTIMAL, cp_model.FEASIBLE) else 0,
         "bound": float(solver.BestObjectiveBound()),
         "status": solver.StatusName(status),
         "members": [i for i in range(n) if status in (cp_model.OPTIMAL, cp_model.FEASIBLE) and solver.Value(x[i])],
